Refactor recon.execute
